revert zipUtils
