+++ src/utils/index.ts
@@ -1,7 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { JSZip } from "./zipUtils";
 export { default as pqUtils } from "./pqUtils";
-export { default as zipUtils } from "./zipUtils";
 export { default as arrayUtils } from "./arrayUtils";
